// src/models/IndexIVFPQModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.12.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Model for configuring an IVFPQ (Inverted File with Product Quantization) index.
 * 
 * Attributes:
 *     type (str): Index type identifier. Defaults to "ivfpq".
 *     pq_dim (int): Dimensionality of PQ codes.
 *     pq_bits (int): Number of bits per quantizer.
 * @export
 * @interface IndexIVFPQModel
 */
export interface IndexIVFPQModel {
    /**
     * 
     * @type {number}
     * @memberof IndexIVFPQModel
     */
    dimension?: number | null;
    /**
     * 
     * @type {string}
     * @memberof IndexIVFPQModel
     */
    type?: string;
    /**
     * 
     * @type {number}
     * @memberof IndexIVFPQModel
     */
    pqDim: number;
    /**
     * 
     * @type {number}
     * @memberof IndexIVFPQModel
     */
    pqBits: number;
}

/**
 * Check if a given object implements the IndexIVFPQModel interface.
 */
export function instanceOfIndexIVFPQModel(value: object): value is IndexIVFPQModel {
    if (!('pqDim' in value) || value['pqDim'] === undefined) return false;
    if (!('pqBits' in value) || value['pqBits'] === undefined) return false;
    return true;
}

export function IndexIVFPQModelFromJSON(json: any): IndexIVFPQModel {
    return IndexIVFPQModelFromJSONTyped(json, false);
}

export function IndexIVFPQModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IndexIVFPQModel {
    if (json == null) {
        return json;
    }
    return {
        
        'dimension': json['dimension'] == null ? undefined : json['dimension'],
        'type': json['type'] == null ? undefined : json['type'],
        'pqDim': json['pq_dim'],
        'pqBits': json['pq_bits'],
    };
}

export function IndexIVFPQModelToJSON(json: any): IndexIVFPQModel {
    return IndexIVFPQModelToJSONTyped(json, false);
}

export function IndexIVFPQModelToJSONTyped(value?: IndexIVFPQModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dimension': value['dimension'],
        'type': value['type'],
        'pq_dim': value['pqDim'],
        'pq_bits': value['pqBits'],
    };
}