// src/models/IndexIVFFlatModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.12.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Model for configuring an IVFFlat (Inverted File with Flat quantization) index.
 * 
 * Attributes:
 *     type (str): Index type identifier. Defaults to "ivfflat".
 * @export
 * @interface IndexIVFFlatModel
 */
export interface IndexIVFFlatModel {
    /**
     * 
     * @type {number}
     * @memberof IndexIVFFlatModel
     */
    dimension?: number | null;
    /**
     * 
     * @type {string}
     * @memberof IndexIVFFlatModel
     */
    type?: string;
}

/**
 * Check if a given object implements the IndexIVFFlatModel interface.
 */
export function instanceOfIndexIVFFlatModel(value: object): value is IndexIVFFlatModel {
    return true;
}

export function IndexIVFFlatModelFromJSON(json: any): IndexIVFFlatModel {
    return IndexIVFFlatModelFromJSONTyped(json, false);
}

export function IndexIVFFlatModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IndexIVFFlatModel {
    if (json == null) {
        return json;
    }
    return {
        
        'dimension': json['dimension'] == null ? undefined : json['dimension'],
        'type': json['type'] == null ? undefined : json['type'],
    };
}

export function IndexIVFFlatModelToJSON(json: any): IndexIVFFlatModel {
    return IndexIVFFlatModelToJSONTyped(json, false);
}

export function IndexIVFFlatModelToJSONTyped(value?: IndexIVFFlatModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dimension': value['dimension'],
        'type': value['type'],
    };
}