// src/models/IndexIVFModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.12.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Model for configuring an IVF (Inverted File) index.
 * 
 * Attributes:
 *     type (str): Index type identifier. Defaults to "ivf".
 * @export
 * @interface IndexIVFModel
 */
export interface IndexIVFModel {
    /**
     * 
     * @type {number}
     * @memberof IndexIVFModel
     */
    dimension?: number | null;
    /**
     * 
     * @type {string}
     * @memberof IndexIVFModel
     */
    type?: string;
}

/**
 * Check if a given object implements the IndexIVFModel interface.
 */
export function instanceOfIndexIVFModel(value: object): value is IndexIVFModel {
    return true;
}

export function IndexIVFModelFromJSON(json: any): IndexIVFModel {
    return IndexIVFModelFromJSONTyped(json, false);
}

export function IndexIVFModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IndexIVFModel {
    if (json == null) {
        return json;
    }
    return {
        
        'dimension': json['dimension'] == null ? undefined : json['dimension'],
        'type': json['type'] == null ? undefined : json['type'],
    };
}

export function IndexIVFModelToJSON(json: any): IndexIVFModel {
    return IndexIVFModelToJSONTyped(json, false);
}

export function IndexIVFModelToJSONTyped(value?: IndexIVFModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dimension': value['dimension'],
        'type': value['type'],
    };
}